requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.28
opencv-contrib-python==4.9.0.80
numpy==1.26.4
pillow==10.2.0
openpyxl==3.1.2
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
cryptography==46.0.7
